issue #162 refactor: directory name change
