+++ src/components/PersonalTodoList/PersonalTodoItem/PersonalTodoItem.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { AiOutlineEdit, AiOutlineDelete } from "react-icons/ai";
 import { useDispatch } from "react-redux";
 
-import BaseCard from "@/components/Base/BaseCard/BaseCard.jsx";
+import BaseCard from "@/components/Ui/BaseCard/BaseCard.jsx";
 import { setEdit } from "@/features/auth/auth-slice.js";
 import { deleteSchedule } from "@/features/schedule/schedule-service.js";
 import { setId } from "@/features/schedule/schedule-slice.js";
